# api-backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IntelliQ API",
        "version": "0.0.1"
    },
    "paths": {
        "/intelliq_api/doanswer/{questionnaireID}/{questionID}/{session}/{optionID}": {
            "post": {
                "summary": "Add an answer to the database",
                "description": "Adds an answer to a question in the database",
                "parameters": [
                    {
                        "in": "path",
                        "name": "questionnaireID",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The questionnaire ID of the answer to add"
                    },
                    {
                        "in": "path",
                        "name": "questionID",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The question ID of the answer to add"
                    },
                    {
                        "in": "path",
                        "name": "session",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The session ID of the answer to add"
                    },
                    {
                        "in": "path",
                        "name": "optionID",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The option ID of the answer to add"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The answer was added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Answer added successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Missing required parameters: questionnaireID, questionID"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intelliq_api/getlastsessid": {
            "get": {
                "summary": "Get last session ID",
                "description": "Returns the lexicographically last session ID",
                "produces": [
                    "application/json",
                    "text/csv"
                ],
                "parameters": [
                    {
                        "name": "format",
                        "description": "Response format (json or csv)",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "json",
                            "csv"
                        ],
                        "default": "json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the last sessid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "sessionID": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Returned if required parameters are missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Returned if no data is found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returned if internal error occurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}